AMP-64212 WIP code clean up - broke down CLI commands into separate classes
