+++ packages/cli/src/actions/BaseAction.ts
@@ -1,0 +1,7 @@
+export interface CliAction {
+  run(options: any): Promise<void>
+}
+
+export abstract class BaseAction implements CliAction {
+  abstract run(options: any): Promise<void>;
+}
